remove(ng): move google_maps out of angular folder
